removed fluentd endpoint from provisioning; output warning if analytics sa not given
