+++ cmd/provision/provision.go
@@ -292,6 +292,9 @@ func (p *provision) run(printf shared.FormatFn) error {
 			if err != nil {
 				return errors.Wrapf(err, "creating analytics secret data")
 			}
+		} else {
+			shared.Errorf("\nWARNING: No analytics service account given via --analytics-sa.")
+			shared.Errorf("\nIMPORTANT: Please make sure the application default credentials where the adapter is run are correctly configured.")
 		}
 	} else {
 		verifyErrors = p.verifyWithoutRetry(config, verbosef)
@@ -387,9 +390,6 @@ func (p *provision) verifyWithRetry(config *server.Config, verbosef shared.Forma
 				shared.Errorf("\nWARNING: Apigee may not be provisioned properly.")
 				shared.Errorf("Unable to verify proxy endpoint(s). Errors:\n")
 				for _, err := range multierr.Errors(verifyErrors) {
-					if strings.Contains(err.Error(), "Unable to get the runtime version") {
-						p.encodeUDCAEndpoint(config, verbosef)
-					}
 					shared.Errorf("  %s", err)
 				}
 				shared.Errorf("\n")
@@ -433,15 +433,6 @@ func (p *provision) verify(config *server.Config, verbosef shared.FormatFn) erro
 
 	verbosef("verifying remote-service proxy...")
 	verifyErrors = multierr.Combine(verifyErrors, p.verifyRemoteServiceProxy(client, verbosef))
-
-	if p.IsGCPManaged {
-		if version, err := p.checkRuntimeVersion(config, client, verbosef); err != nil {
-			err = errors.Wrapf(err, "Unable to get the runtime version")
-			verifyErrors = multierr.Combine(verifyErrors, err)
-		} else if version >= "1.3.0" || version == "unknown" {
-			p.encodeUDCAEndpoint(config, verbosef)
-		}
-	}
 
 	return verifyErrors
 }
